Adding simple logging view
The most significant changes in the code are related to the implementation of a new logging mechanism using the `Serilog` library and the creation of a separate log view. The `Serilog` library has been added to the project and a new `ILogger` property named `Logger` has been added to the `App` class for logging purposes. The `logViewModel` and `logPage` have been added as transient services, and the `App_UnhandledException` method now logs exceptions using the `Logger` property. The `ItemsRepeaterLogBroker` has been removed, suggesting a change in the logging mechanism.

A new `logViewModel` class has been added to manage the state and operations of the log view. The `MainPage.xaml` file has been significantly modified, with many elements related to the log viewer being removed, indicating that the log viewer has been moved to a separate page. A new `MenuViewslogCommand` has been added to the `ShellViewModel` class for navigation to the log view.

The `logPage.xaml` and `logPage.xaml.cs` files have been added to define the log view and its behavior. New resources for the text of the menu item that navigates to the log view and the title of the log view have been added to `Resources.resw`. The `PageService` class has been updated to map the `logViewModel` to the `logPage`, and the `ShellPage.xaml` file has been updated to include a menu item that executes the `MenuViewslogCommand` when clicked.

Changes:
1. Addition of `Serilog` library for logging (App.xaml.cs).
2. Addition of `ILogger` property `Logger` to `App` class (App.xaml.cs).
3. Addition of `logViewModel` and `logPage` as transient services (App.xaml.cs).
4. Logging of exceptions in `App_UnhandledException` method (App.xaml.cs).
5. Removal of `ItemsRepeaterLogBroker` (MainWindow.xaml.cs, MainPage.xaml.cs).
6. Addition of `logViewModel` class (logViewModel.cs).
7. Significant modifications to `MainPage.xaml`.
8. Addition of `MenuViewslogCommand` to `ShellViewModel` class (ShellViewModel.cs).
9. Addition of `logPage.xaml` and `logPage.xaml.cs` files.
10. Addition of new resources to `Resources.resw`.
11. Update to `PageService` class to map `logViewModel` to `logPage` (PageService.cs).
12. Update to `ShellPage.xaml` to include new menu item.